Plotting (ACHMC)

Atomic CHMC

MarkovWeightedEFMs.Plots.plot_mapped_reaction — Function

function plot_mapped_reaction(#
    s::String,
    fname::String = "";
    view::Bool = false,
    canvas_width::Int64 = 3000,
    canvas_height::Int64 = 1000

)

Plot mapped reaction SMILES string s as an SVG and save to fname if specified.

canvas_width is the width of the SVG.

canvas_height is the height of the SVG.

view=true will plot the SVG assuming a plotting backend is specified. For example, loading the ElectronDisplay package will plot the SVG to an Electron window.

MarkovWeightedEFMs.Plots.plot_atomic_chmc — Function

function plot_atomic_chmc(#
    res::CHMCAtomicSummary,
    S::Matrix{Int16},
    mets::Vector{String},
    rs::Vector{String};
    node_label_textsize::Real=15,
    edge_label_textsize::Real=12,
    arrow_shift::Real=0.85,
    x_pad::Real=0.75,
    y_pad::Real=0.75,
    tfactor::Real=0.15,
    tangents::Tuple{#
        Tuple{<:Real, <:Real},
        Tuple{<:Real, <:Real}
    } = ((1,0),(0,1)),
    show_all::Bool=false,
    width::Int64=620,
    height::Int64=310

)

Plot atomic cycle-history Markov chain.

S is the m by n stoichiometry matrix.

mets is the vector of metabolite names of length m.

rs is the vector of reaction SMILES strings of length n.

node_label_textsize is the text size of the node labels indexed from T.

edge_label_textsize is the text size of the edge labels taken from T.

arrow_shift is the percentage shift of the arrow head from src to dst.

x_pad is the left/right x coordinate padding of the plotting box.

y_pad is the up/down y coordinate padding of the plotting box.

tfactor scales the distance of the bezier control point relative to the distance of the src and dst nodes.

tangents is the tangent of the src vertex and dst vertex.

show_all=true explicitly plots the upstream transition from all EFMs that pass through the initial state/node I. By default, these arrows stemming from the green nodes are omitted for visual clarity.

width and height specify the plotting window dimensions in pixel units.

Index