Functions

MarkovWeightedEFMs.reshape_efm_matrix — Function

reshape_efm_matrix(#
    ϕ::Matrix{Int64},
    S::Matrix{<:Real}
)

Convert a matrix of EFMs ϕ to a nested vector of EFMs from a stoichiometry matrix S. Stoichiometry matrix may only contain unimolecular reactions.

ϕ is the n by k EFM matrix with n reactions (rows) and k EFMs (cols).

S is the m by n stoichiometry matrix with m metabolites (rows) and n reactions (cols).

Examples

julia> ϕ = [#
  1  1  0  0  0  0
  1  0  1  0  0  0
  0  0  1  0  0  1
  0  1  0  0  0  1
  1  0  0  1  0  0
  0  0  0  1  0  1
  0  1  0  0  0  1
  1  1  0  0  0  0
  0  0  0  0  1  0
  0  0  0  0  1  0
  1  1  0  0  0  0
]
julia> S = [#
  -1  0  0  0  0  0  0  0  0  0  1
   1 -1  1 -1  0  0  0  0  0  0  0
   0  1 -1  0 -1  1  0  0  0  0  0
   0  0  0  1  0  0 -1  0  0  0  0
   0  0  0  0  1 -1  1 -1 -1  1  0
   0  0  0  0  0  0  0  1  0  0 -1
   0  0  0  0  0  0  0  0  1 -1  0
]
julia> efm_vector = reshape_efm_matrix(ϕ, S)
6-element Vector{Vector{Int64}}:
 [1, 2, 3, 5, 6, 1]
 [1, 2, 4, 5, 6, 1]
 [2, 3, 2]
 [3, 5, 3]
 [5, 7, 5]
 [3, 2, 4, 5, 3]

MarkovWeightedEFMs.reshape_efm_vector — Function

reshape_efm_vector(#
    ϕ::Vector{Vector{Int64}},
    S::Matrix{<:Real}
)

Convert nested vector of EFM indices ϕ with length k to an n by k matrix of EFMs based on m by n stoichiometry matrix S. Stoichiometry matrix may only contain unimolecular reactions.

ϕ is the nested vector of EFMs with length k and elements corresponding to EFM metabolite indices in S.

S is the m by n stoichiometry matrix with m metabolites (rows) and n reactions (cols).

Examples

julia> ϕ = [#
  [1, 2, 3, 5, 6, 1],
  [1, 2, 4, 5, 6, 1],
  [2, 3, 2], [3, 5, 3], [5, 7, 5], [2, 4, 5, 3, 2]
]
julia> S = [#
  -1  0  0  0  0  0  0  0  0  0  1
   1 -1  1 -1  0  0  0  0  0  0  0
   0  1 -1  0 -1  1  0  0  0  0  0
   0  0  0  1  0  0 -1  0  0  0  0
   0  0  0  0  1 -1  1 -1 -1  1  0
   0  0  0  0  0  0  0  1  0  0 -1
   0  0  0  0  0  0  0  0  1 -1  0
]
julia> efm_matrix = reshape_efm_vector(ϕ, S)
11x6 Matrix{Int64}:
 1  1  0  0  0  0
 1  0  1  0  0  0
 0  0  1  0  0  1
 0  1  0  0  0  1
 1  0  0  1  0  0
 0  0  0  1  0  1
 0  1  0  0  0  1
 1  1  0  0  0  0
 0  0  0  0  1  0
 0  0  0  0  1  0
 1  1  0  0  0  0

MarkovWeightedEFMs.steady_state_efm_distribution — Function

steady_state_efm_distribution(#
    S::Matrix{<:Int64},
    v::Vector{<:Real},
    I::Int64=1
)

Enumerates the EFMs from stoichiometry matrix S and compute the steady state probabilities of each EFM according to the cycle-history, discrete-time Markov chain model.

S is a fully-connected, unimolecular, m by n stoichiometry matrix with m metabolites and n reactions.

v is the n-length steady state flux vector associated with S.

I is the initial starting state for rooting the cycle-history Markov chain. The choice of initial starting state does not affect the steady state EFM probabilities. The default is 1 and must be a whole number between 1:m.

Example

julia> S = [#
 -1  0  0  0  0  0  0  0  0  0  1
  1 -1  1 -1  0  0  0  0  0  0  0
  0  1 -1  0 -1  1  0  0  0  0  0
  0  0  0  1  0  0 -1  0  0  0  0
  0  0  0  0  1 -1  1 -1  1 -1  0
  0  0  0  0  0  0  0  0  0  1 -1
  0  0  0  0  0  0  0  1 -1  0  0
];
julia> v = [3, 2, 1, 2, 3, 2, 2, 1, 1, 3, 3];
julia> res = steady_state_efm_distribution(S, v);
julia> res.e # EFM state sequences
6-element Vector{Vector{Int64}}:
 [3, 2, 3]
 [5, 3, 5]
 [7, 5, 7]
 [6, 1, 2, 4, 5, 6]
 [3, 2, 4, 5, 3]
 [6, 1, 2, 3, 5, 6]

julia> res.p # EFM probabilities
6-element Vector{Float64}:
 0.10638297872340426
 0.25531914893617025
 0.14893617021276595
 0.25531914893617025
 0.0425531914893617
 0.1914893617021277

julia> res.w # EFM weights
6-element Vector{Float64}:
 0.7142857142857142
 1.7142857142857144
 0.9999999999999999
 1.7142857142857144
 0.2857142857142857
 1.2857142857142858

MarkovWeightedEFMs.stoich_to_transition — Function

stoich_to_transition(S::Matrix{<:Real}, v::Vector{<:Real})

Convert a stoichiometry matrix with vector of steady state fluxes to a right stochastic transition probability matrix with rows summing to one.

S is the m by n stoichiometry matrix with m metabolites and n reactions.

v is the steady state flux vector with length n.

Examples

julia> S = [#
  -1  0  0  0  0  0  0  0  0  0  1
   1 -1  1 -1  0  0  0  0  0  0  0
   0  1 -1  0 -1  1  0  0  0  0  0
   0  0  0  1  0  0 -1  0  0  0  0
   0  0  0  0  1 -1  1 -1 -1  1  0
   0  0  0  0  0  0  0  1  0  0 -1
   0  0  0  0  0  0  0  0  1 -1  0
]
julia> v = [2, 2, 2, 2, 2, 2, 2, 2, 4]
julia> stoich_to_transition(S, v)
7x7 Matrix{Float64}:
 0.0  1.0   0.0       0.0  0.0   0.0  0.0
 0.0  0.0   0.5       0.5  0.0   0.0  0.0
 0.0  0.25  0.0       0.0  0.75  0.0  0.0
 0.0  0.0   0.0       0.0  1.0   0.0  0.0
 0.0  0.0   0.333333  0.0  0.0   0.5 0.166667
 1.0  0.0   0.0       0.0  0.0   0.0  0.0
 0.0  0.0   0.0       0.0  1.0   0.0  0.0

MarkovWeightedEFMs.tree_plot — Function

function tree_plot(#
    T::Matrix{<:Real},
    I::Int64=1;
    node_label_textsize::Real=15,
    edge_label_textsize::Real=12,
    arrow_shift::Real=0.85,
    x_pad::Real=0.75,
    y_pad::Real=0.75,
    tfactor::Real=0.15,
    tangents::Tuple{#
        Tuple{<:Real, <:Real},
        Tuple{<:Real, <:Real}
    } = ((1,0),(0,1)),
    show_all::Bool=false

)

Plot augmented prefix tree from transition probability matrix T rooted on state/node I.

T is a right stochastic transition probability matrix. Alternatively, one could specify a right generator matrix (with rows summing to zero).

I is a state in 1:size(T,1).

nodelabeltextsize is the text size of the node labels indexed from T.

edgelabeltextsize is the text size of the edge labels taken from T.

arrow_shift is the percentage shift of the arrow head from src to dst.

x_pad is the left/right x coordinate padding of the plotting box.

y_pad is the up/down y coordinate padding of the plotting box.

tfactor scales the distance of the bezier control point relative to the distance of the src and dst nodes.

tangents is the tangent of the src vertex and dst vertex.

show_all = true explicitly plots the upstream transition from all EFMs that pass through the initial state/node I. By default, these arrows stemming from the green nodes are omitted for visual clarity.

Index