Plotting

CHMC

MarkovWeightedEFMs.Plots.plot_chmc — Function

function plot_chmc(#
    T::Matrix{<:Real},
    I::Int64=1;
    node_label_textsize::Real=15,
    edge_label_textsize::Real=12,
    arrow_shift::Real=0.85,
    x_pad::Real=0.75,
    y_pad::Real=0.75,
    tfactor::Real=0.15,
    tangents::Tuple{#
        Tuple{<:Real, <:Real},
        Tuple{<:Real, <:Real}
    } = ((1,0),(0,1)),
    show_all::Bool=false

)

Plot cycle-history Markov chain from transition probability matrix T rooted on state/node I.

T is a right stochastic transition probability matrix with rows summing to one.

I is a state in 1:size(T,1).

node_label_textsize is the text size of the node labels indexed from T.

edge_label_textsize is the text size of the edge labels taken from T.

arrow_shift is the percentage shift of the arrow head from src to dst.

x_pad is the left/right x coordinate padding of the plotting box.

y_pad is the up/down y coordinate padding of the plotting box.

tfactor scales the distance of the bezier control point relative to the distance of the src and dst nodes.

tangents is the tangent of the src vertex and dst vertex.

show_all=true explicitly plots the upstream transition from all EFMs that pass through the initial state/node I. By default, these arrows stemming from the green nodes are omitted for visual clarity.

Index

MarkovWeightedEFMs.Plots.plot_chmc